# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
  agencies.Agency:
    properties:
      address:
        type: string
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      updatedAt:
        type: string
    type: object
...
  agencies.WriteAgencyRequest:
    properties:
      address:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  attendances.AttendEnum:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - Absented
    - Attended
    - Excused
    - Dayoff
    - Holiday
  attendances.Attendance:
    properties:
      attendedAt:
        type: string
      attendedStatus:
        enum:
        - absented
        - attended
        - excused
        - dayoff
        - holiday
        type: string
      classId:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      studentId:
        type: integer
      updatedAt:
        type: string
    type: object
  attendances.CreateAttendanceRequest:
    properties:
      attendedAt:
        type: string
      attendedStatus:
        enum:
        - absented
        - attended
        - excused
        - dayoff
        - holiday
        type: string
      classId:
        type: integer
      studentId:
        type: integer
    type: object
  attendances.UpdateAttendanceRequest:
    properties:
      attendedStatus:
        $ref: '#/definitions/attendances.AttendEnum'
      id:
        type: integer
    type: object
  classes.Class:
    properties:
      agencyId:
        type: integer
      createdAt:
        type: string
      grade:
        description: |-
          Grade type:
          * buds - Children who is 3 yo.
          * seed - Children who is 4 yo.
          * leaf - Children who is 5 yo.
          * toddler - Children who is lower than 2 yo.
        enum:
        - buds
        - seed
        - leaf
        - ' toddler'
        type: string
      id:
        type: integer
      name:
        type: string
      teacherId:
        type: string
      updatedAt:
        type: string
    type: object
...
  classes.WriteClassRequest:
    properties:
      agencyId:
        type: integer
      grade:
        description: |-
          Grade type:
          * buds - Children who is 3 yo.
          * seed - Children who is 4 yo.
          * leaf - Children who is 5 yo.
          * toddler - Children who is lower than 2 yo.
        enum:
        - buds
        - seed
        - leaf
        - ' toddler'
        type: string
      name:
        type: string
      teacherId:
        type: string
    type: object
...
  registrations.FindRegistrationResp:
    properties:
      data:
        items:
          $ref: '#/definitions/registrations.Registration'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
    type: object
  registrations.Registration:
    properties:
      agencyId:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      isProcessed:
        type: boolean
      note:
        type: string
      parentName:
        type: string
      phoneNumber:
        type: string
      studentClass:
        description: |-
          Class type:
          * buds - Children who is 3 yo.
          * seed - Children who is 4 yo.
          * leaf - Children who is 5 yo.
          * toddler - Children who is 1 - 3 yo.
        enum:
        - buds
        - seed
        - leaf
        - ' toddler'
        type: string
      studentDob:
        type: string
      studentName:
        type: string
      updatedAt:
        type: string
    type: object
  registrations.WriteRegistrationRequest:
    properties:
      agencyId:
        type: integer
      note:
        type: string
      parentName:
        type: string
      phoneNumber:
        type: string
      studentClass:
        description: |-
          Class type:
          * buds - Children who is 3 yo.
          * seed - Children who is 4 yo.
          * leaf - Children who is 5 yo.
          * toddler - Children who is 1 - 3 yo.
        enum:
        - buds
        - seed
        - leaf
        - ' toddler'
        type: string
      studentDob:
        type: string
      studentName:
        type: string
    type: object
  roles.FindRoleResp:
    properties:
      data:
        items:
          $ref: '#/definitions/roles.Role'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
    type: object
  roles.Role:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  roles.WriteRoleRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
...
  /agencies/{id}:
    get:
      consumes:
      - application/json
      description: Get one agency
      parameters:
      - description: Agency ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agencies.Agency'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get agency details api
      tags:
      - Agency
    put:
      consumes:
      - application/json
      description: Update agency
      parameters:
      - description: Update agency body
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/agencies.WriteAgencyRequest'
      - description: Agency ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agencies.Agency'
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update agency api
      tags:
      - Agency
  /attendances:
    get:
      consumes:
      - application/json
      description: Get list attendance
      parameters:
      - description: Class id
        in: query
        name: classId
        required: true
        type: string
      - description: Time range
        in: query
        name: period
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/attendances.Attendance'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get list attendance details api
      tags:
      - Attendance
    patch:
      consumes:
      - application/json
      description: Patch attendance
      parameters:
      - description: Patch attendance body
        in: body
        name: post
        required: true
        schema:
          items:
            $ref: '#/definitions/attendances.UpdateAttendanceRequest'
          type: array
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/attendances.Attendance'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Patch attendance api
      tags:
      - Attendance
    post:
      consumes:
      - application/json
      description: Insert attendance
      parameters:
      - description: Create attendance body
        in: body
        name: post
        required: true
        schema:
          items:
            $ref: '#/definitions/attendances.CreateAttendanceRequest'
          type: array
      responses:
        "200":
          description: OK
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create attendance api
      tags:
      - Attendance
...
  /classes/{id}:
    get:
      consumes:
      - application/json
      description: Get one class
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/classes.Class'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get class details api
      tags:
      - Class
    put:
      consumes:
      - application/json
      description: Update class
      parameters:
      - description: Update class body
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/classes.WriteClassRequest'
      - description: Class ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/classes.Class'
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update class api
      tags:
      - Class
  /registrations:
    delete:
      consumes:
      - application/json
      description: Delete registration
      parameters:
      - collectionFormat: csv
        description: Registration IDs
        in: query
        items:
          type: integer
        name: ids
        required: true
        type: array
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete registration api
      tags:
      - Registration
    get:
      consumes:
      - application/json
      description: Get list registration
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: pageSize
        type: integer
      - default: desc
        description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/registrations.FindRegistrationResp'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get list registration details api
      tags:
      - Registration
    post:
      consumes:
      - application/json
      description: Insert registration
      parameters:
      - description: Create registration body
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/registrations.WriteRegistrationRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/registrations.Registration'
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create registration api
      tags:
      - Registration
...
  /roles:
    get:
      consumes:
      - application/json
      description: Get list role
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: pageSize
        type: integer
      - default: desc
        description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roles.FindRoleResp'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get list role details api
      tags:
      - Role
    post:
      consumes:
      - application/json
      description: Insert role
      parameters:
      - description: Create role body
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/roles.WriteRoleRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roles.Role'
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create role api
      tags:
      - Role
  /roles/{id}:
    get:
      consumes:
      - application/json
      description: Get one role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roles.Role'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get role details api
      tags:
      - Role
    put:
      consumes:
      - application/json
      description: Update role
      parameters:
      - description: Update role body
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/roles.WriteRoleRequest'
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roles.Role'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update role api
      tags:
      - Role
...
securityDefinitions:
  ApiKeyAuth:
    description: Apply "bearer " before token in authorization
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"